+++ test/once_connection/models/common.rs
@@ -1,0 +1,2 @@
+/* This file is generated and managed by dsync */
+type Connection = diesel::r2d2::PooledConnection<diesel::r2d2::ConnectionManager<diesel::PgConnection>>;