+++ test/once_connection_once_common_structs/models/common.rs
@@ -1,0 +1,16 @@
+/* This file is generated and managed by dsync */
+/// Result of a `.paginate` function
+#[derive(Debug, Serialize)]
+pub struct PaginationResult<T> {
+    /// Result items from the current page
+    pub items: Vec<T>,
+    /// Count of how many items there are in total
+    pub total_items: i64,
+    /// Current page, 0-based index
+    pub page: i64,
+    /// Size of a page
+    pub page_size: i64,
+    /// Number of pages in total
+    pub num_pages: i64,
+}
+type Connection = diesel::r2d2::PooledConnection<diesel::r2d2::ConnectionManager<diesel::PgConnection>>;